Cleanup and move code
